refactor: 💡 merge model and collection abilities
