--- addons/api/app/abilities/collection.js
@@ -1,1 +1,0 @@
-export { default } from 'api/abilities/collection';
