+++ addons/api/tests/unit/abilities/model-test.js
@@ -65,4 +65,32 @@ module('Unit | Abilities | Model', function (hooks) {
     model.authorized_actions = [];
     assert.notOk(service.can('delete model', model));
   });
+
+  // =collections
+
+  test('it reflects when a resource may be listed based on authorized_collection_actions', function (assert) {
+    assert.expect(2);
+    const service = this.owner.lookup('service:can');
+    const model = {
+      authorized_collection_actions: {
+        foobars: ['list'],
+      },
+    };
+    assert.ok(service.can('list model', model, { collection: 'foobars' }));
+    model.authorized_collection_actions.foobars = [];
+    assert.notOk(service.can('list model', model, { collection: 'foobars' }));
+  });
+
+  test('it reflects when a resource may be created based on authorized_collection_actions', function (assert) {
+    assert.expect(2);
+    const service = this.owner.lookup('service:can');
+    const model = {
+      authorized_collection_actions: {
+        foobars: ['create'],
+      },
+    };
+    assert.ok(service.can('create model', model, { collection: 'foobars' }));
+    model.authorized_collection_actions.foobars = [];
+    assert.notOk(service.can('create model', model, { collection: 'foobars' }));
+  });
 });
