+++ ui/admin/app/routes/scopes/scope/credential-stores.js
@@ -21,7 +21,7 @@ export default class ScopesScopeCredentialStoresRoute extends Route {
     const scope = this.modelFor('scopes.scope');
     const { id: scope_id } = scope;
     if (
-      this.can.can('list collection', scope, {
+      this.can.can('list model', scope, {
         collection: 'credential-stores',
       })
     ) {
