+++ ui/admin/app/routes/scopes/scope/host-catalogs/host-catalog/hosts.js
@@ -25,7 +25,7 @@ export default class ScopesScopeHostCatalogsHostCatalogHostsRoute extends Route 
     );
     const { id: host_catalog_id } = hostCatalog;
     if (
-      this.can.can('list collection', hostCatalog, {
+      this.can.can('list model', hostCatalog, {
         collection: 'hosts',
       })
     ) {
