+++ ui/admin/app/templates/scopes/scope/credential-stores/index.hbs
@@ -15,7 +15,7 @@
 
   <page.actions>
 
-    {{#if (can 'create collection' this.scope collection='credential-stores')}}
+    {{#if (can 'create model' this.scope collection='credential-stores')}}
       {{#if @model}}
         <Rose::LinkButton
           @route='scopes.scope.credential-stores.new'
@@ -82,13 +82,11 @@
           as |message|
         >
           <message.description>
-            {{#if
-              (can 'list collection' this.scope collection='credential-stores')
-            }}
+            {{#if (can 'list model' this.scope collection='credential-stores')}}
               {{! can list (at least):  show default welcome message}}
               {{t 'resources.credential-store.description'}}
             {{else if (can
-              'create collection' this.scope collection='credential-stores'
+              'create model' this.scope collection='credential-stores'
             )}}
               {{! can create (only):  show create-but-not-list welcome message}}
               {{t
@@ -103,9 +101,7 @@
               }}
             {{/if}}
           </message.description>
-          {{#if
-            (can 'create collection' this.scope collection='credential-stores')
-          }}
+          {{#if (can 'create model' this.scope collection='credential-stores')}}
             <message.link @route='scopes.scope.credential-stores.new'>
               <Rose::Icon @name='flight-icons/svg/plus-circle-16' />
               {{t 'titles.new'}}
