+++ ui/admin/app/templates/scopes/scope/groups/index.hbs
@@ -13,7 +13,7 @@
   </page.header>
 
   <page.actions>
-    {{#if (can 'create collection' this.scope collection='groups')}}
+    {{#if (can 'create model' this.scope collection='groups')}}
       {{#if @model}}
         <Rose::LinkButton @route='scopes.scope.groups.new' @style='primary'>{{t
             'titles.new'
@@ -71,10 +71,10 @@
           as |message|
         >
           <message.description>
-            {{#if (can 'list collection' this.scope collection='groups')}}
+            {{#if (can 'list model' this.scope collection='groups')}}
               {{! can list (at least):  show default welcome message}}
               {{t 'resources.group.description'}}
-            {{else if (can 'create collection' this.scope collection='groups')}}
+            {{else if (can 'create model' this.scope collection='groups')}}
               {{! can create (only):  show create-but-not-list welcome message}}
               {{t
                 'descriptions.create-but-not-list'
@@ -88,7 +88,7 @@
               }}
             {{/if}}
           </message.description>
-          {{#if (can 'create collection' this.scope collection='groups')}}
+          {{#if (can 'create model' this.scope collection='groups')}}
             <message.link @route='scopes.scope.groups.new'>
               <Rose::Icon @name='flight-icons/svg/plus-circle-16' />
               {{t 'titles.new'}}
