+++ ui/admin/app/templates/scopes/scope/host-catalogs/host-catalog/-navigation.hbs
@@ -2,12 +2,12 @@
   <nav.link @route='scopes.scope.host-catalogs.host-catalog.index'>
     {{t 'titles.details'}}
   </nav.link>
-  {{#if (can 'navigate collection' @model collection='hosts')}}
+  {{#if (can 'navigate model' @model collection='hosts')}}
     <nav.link @route='scopes.scope.host-catalogs.host-catalog.hosts'>
       {{t 'resources.host.title_plural'}}
     </nav.link>
   {{/if}}
-  {{#if (can 'navigate collection' @model collection='host-sets')}}
+  {{#if (can 'navigate model' @model collection='host-sets')}}
     <nav.link @route='scopes.scope.host-catalogs.host-catalog.host-sets'>
       {{t 'resources.host-set.title_plural'}}
     </nav.link>
