+++ ui/admin/app/templates/scopes/scope/host-catalogs/host-catalog/hosts/index.hbs
@@ -69,13 +69,11 @@
     >
       <message.description>
 
-        {{#if (can 'list collection' this.hostCatalog collection='hosts')}}
+        {{#if (can 'list model' this.hostCatalog collection='hosts')}}
           {{! can list (at least):  show default welcome message}}
           {{t 'resources.host.description'}}
 
-        {{else if (can
-          'create collection' this.hostCatalog collection='hosts'
-        )}}
+        {{else if (can 'create model' this.hostCatalog collection='hosts')}}
           {{! can create (only):  show create-but-not-list welcome message}}
           {{t
             'descriptions.create-but-not-list'
@@ -90,7 +88,7 @@
         {{/if}}
       </message.description>
 
-      {{#if (can 'create collection' this.hostCatalog collection='hosts')}}
+      {{#if (can 'create model' this.hostCatalog collection='hosts')}}
 
         <message.link
           @route='scopes.scope.host-catalogs.host-catalog.hosts.new'
