+++ ui/admin/app/templates/scopes/scope/targets/index.hbs
@@ -15,7 +15,7 @@
   </page.header>
 
   <page.actions>
-    {{#if (can 'create collection' this.scope collection='targets')}}
+    {{#if (can 'create model' this.scope collection='targets')}}
       {{#if @model}}
         <Rose::LinkButton @route='scopes.scope.targets.new' @style='primary'>{{t
             'titles.new'
@@ -81,12 +81,10 @@
         >
           <message.description>
 
-            {{#if (can 'list collection' this.scope collection='targets')}}
+            {{#if (can 'list model' this.scope collection='targets')}}
               {{! can list (at least):  show default welcome message}}
               {{t 'resources.target.description'}}
-            {{else if (can
-              'create collection' this.scope collection='targets'
-            )}}
+            {{else if (can 'create model' this.scope collection='targets')}}
               {{! can create (only):  show create-but-not-list welcome message}}
               {{t
                 'descriptions.create-but-not-list'
@@ -102,7 +100,7 @@
             {{/if}}
           </message.description>
 
-          {{#if (can 'create collection' this.scope collection='targets')}}
+          {{#if (can 'create model' this.scope collection='targets')}}
             <message.link @route='scopes.scope.targets.new'>
               <Rose::Icon @name='flight-icons/svg/plus-circle-16' />
               {{t 'titles.new'}}
