+++ ui/admin/app/templates/scopes/scope/users/index.hbs
@@ -15,7 +15,7 @@
   </page.header>
 
   <page.actions>
-    {{#if (can 'create collection' this.scope collection='users')}}
+    {{#if (can 'create model' this.scope collection='users')}}
       {{#if @model}}
         <Rose::LinkButton @route='scopes.scope.users.new' @style='primary'>{{t
             'titles.new'
@@ -73,10 +73,10 @@
           as |message|
         >
           <message.description>
-            {{#if (can 'list collection' this.scope collection='users')}}
+            {{#if (can 'list model' this.scope collection='users')}}
               {{! can list (at least):  show default welcome message}}
               {{t 'resources.user.description'}}
-            {{else if (can 'create collection' this.scope collection='users')}}
+            {{else if (can 'create model' this.scope collection='users')}}
               {{! can create (only):  show create-but-not-list welcome message}}
               {{t
                 'descriptions.create-but-not-list'
@@ -90,7 +90,7 @@
               }}
             {{/if}}
           </message.description>
-          {{#if (can 'create collection' this.scope collection='users')}}
+          {{#if (can 'create model' this.scope collection='users')}}
             <message.link @route='scopes.scope.users.new'>
               <Rose::Icon @name='flight-icons/svg/plus-circle-16' />
               {{t 'titles.new'}}
