+++ ui/admin/tests/unit/abilities/collection-test.js
@@ -1,7 +1,7 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 
-module('Unit | Abilities | Collection', function (hooks) {
+module('Unit | Abilities | model', function (hooks) {
   setupTest(hooks);
 
   test('it reflects when a resource may be navigated to based on list and create actions', function (assert) {
@@ -11,19 +11,13 @@ module('Unit | Abilities | Collection', function (hooks) {
       authorized_collection_actions: { foobars: [] },
     };
     assert.notOk(
-      service.can('navigate collection', model, { collection: 'foobars' })
+      service.can('navigate model', model, { collection: 'foobars' })
     );
     model.authorized_collection_actions.foobars = ['list'];
-    assert.ok(
-      service.can('navigate collection', model, { collection: 'foobars' })
-    );
+    assert.ok(service.can('navigate model', model, { collection: 'foobars' }));
     model.authorized_collection_actions.foobars = ['create'];
-    assert.ok(
-      service.can('navigate collection', model, { collection: 'foobars' })
-    );
+    assert.ok(service.can('navigate model', model, { collection: 'foobars' }));
     model.authorized_collection_actions.foobars = ['list', 'create'];
-    assert.ok(
-      service.can('navigate collection', model, { collection: 'foobars' })
-    );
+    assert.ok(service.can('navigate model', model, { collection: 'foobars' }));
   });
 });
