+++ ui/desktop/app/abilities/model.js
@@ -1,3 +1,12 @@
 import ModelAbility from 'api/abilities/model';
 
-export default class OverrideModelAbility extends ModelAbility {}
+export default class OverrideModelAbility extends ModelAbility {
+  /**
+   * Navigating to a resource is allowed if either list or create grants
+   * are present.
+   * @type {boolean}
+   */
+  get canNavigate() {
+    return this.canList || this.canCreate;
+  }
+}
